requirements.txt
aiohttp==3.6.2
aiosqlite==0.15.0
async-timeout==3.0.1
attrs==20.2.0
chardet==3.0.4
configobj==5.0.6
databases==0.3.2
discord.py==1.5.0
idna==2.10
lxml==4.5.2
module-wrapper==0.2.4
mpmath==1.1.0
multidict==4.7.6
poetry-version==0.1.5
six==1.15.0
SQLAlchemy==1.3.19
sympy==1.6
tomlkit==0.5.11
typing-extensions==3.7.4.3
websockets==6.0
yarl==1.6.0
